create more idea str utilities
