+++ src/a17_idea_logic/idea_db_tool.py
@@ -103,18 +103,6 @@ def get_relevant_columns_dataframe(
     return src_df[relevant_cols_in_order]
 
 
-def yell_raw_str() -> str:
-    return "yell_raw"
-
-
-def yell_agg_str() -> str:
-    return "yell_agg"
-
-
-def yell_valid_str() -> str:
-    return "yell_valid"
-
-
 def get_yell_raw_grouping_with_all_values_equal_df(
     x_df: DataFrame, groupby_list: list, idea_number: str
 ) -> DataFrame:
