+++ src/a18_etl_toolbox/test_tran/test_otx_event_idea_sheets2inx_event_idea_sheets.py
@@ -9,11 +9,8 @@ from src.a06_bud_logic._utils.str_a06 import (
 )
 from src.a16_pidgin_logic._utils.str_a16 import pidgin_filename
 from src.a16_pidgin_logic.pidgin import pidginunit_shop
-from src.a17_idea_logic.idea_db_tool import (
-    upsert_sheet,
-    yell_valid_str,
-    sheet_exists,
-)
+from src.a17_idea_logic._utils.str_a17 import yell_valid_str
+from src.a17_idea_logic.idea_db_tool import upsert_sheet, sheet_exists
 from src.a18_etl_toolbox.transformers import (
     etl_otz_event_ideas_to_inz_events,
     get_most_recent_event_int,
