+++ src/a18_etl_toolbox/test_tran/test_otz_face_ideas_sheets2otz_event_ideas_sheets.py
@@ -2,11 +2,8 @@ from src.a00_data_toolbox.file_toolbox import create_path
 from src.a02_finance_logic._utils.strs_a02 import fisc_tag_str
 from src.a06_bud_logic._utils.str_a06 import face_name_str, event_int_str
 from src.a15_fisc_logic._utils.str_a15 import cumlative_minute_str, hour_tag_str
-from src.a17_idea_logic.idea_db_tool import (
-    upsert_sheet,
-    yell_valid_str,
-    sheet_exists,
-)
+from src.a17_idea_logic._utils.str_a17 import yell_valid_str
+from src.a17_idea_logic.idea_db_tool import upsert_sheet, sheet_exists
 from src.a18_etl_toolbox.transformers import etl_otz_face_ideas_to_otz_event_otx_ideas
 from src.a18_etl_toolbox._utils.env_a18 import (
     get_module_temp_dir,
