+++ src/a18_etl_toolbox/test_tran/test_yell_agg_sheets2_yell_valid_sheets.py
@@ -2,12 +2,8 @@ from src.a00_data_toolbox.file_toolbox import create_path
 from src.a02_finance_logic._utils.strs_a02 import fisc_tag_str
 from src.a06_bud_logic._utils.str_a06 import face_name_str, event_int_str
 from src.a15_fisc_logic._utils.str_a15 import cumlative_minute_str, hour_tag_str
-from src.a17_idea_logic.idea_db_tool import (
-    sheet_exists,
-    upsert_sheet,
-    yell_agg_str,
-    yell_valid_str,
-)
+from src.a17_idea_logic._utils.str_a17 import yell_valid_str, yell_agg_str
+from src.a17_idea_logic.idea_db_tool import sheet_exists, upsert_sheet
 from src.a18_etl_toolbox.transformers import (
     etl_yell_agg_non_pidgin_ideas_to_yell_valid,
 )
