+++ src/a18_etl_toolbox/test_tran/test_yell_agg_sheets2pidgin_acct_raw_sheets.py
@@ -8,11 +8,8 @@ from src.a16_pidgin_logic._utils.str_a16 import (
     otx_name_str,
     unknown_word_str,
 )
-from src.a17_idea_logic.idea_db_tool import (
-    get_sheet_names,
-    upsert_sheet,
-    yell_agg_str,
-)
+from src.a17_idea_logic._utils.str_a17 import yell_agg_str
+from src.a17_idea_logic.idea_db_tool import get_sheet_names, upsert_sheet
 from src.a18_etl_toolbox.tran_path import create_yell_pidgin_path
 from src.a18_etl_toolbox.pidgin_agg import PidginPrimeColumns
 from src.a18_etl_toolbox.transformers import etl_yell_agg_to_pidgin_name_raw
