+++ src/a18_etl_toolbox/test_tran/test_yell_agg_sheets2yell_events_sheets.py
@@ -7,8 +7,8 @@ from src.a00_data_toolbox.db_toolbox import (
 from src.a02_finance_logic._utils.strs_a02 import fisc_tag_str
 from src.a06_bud_logic._utils.str_a06 import face_name_str, event_int_str
 from src.a15_fisc_logic._utils.str_a15 import cumlative_minute_str, hour_tag_str
-from src.a17_idea_logic._utils.str_a17 import idea_number_str
-from src.a17_idea_logic.idea_db_tool import yell_agg_str, create_idea_sorted_table
+from src.a17_idea_logic._utils.str_a17 import idea_number_str, yell_agg_str
+from src.a17_idea_logic.idea_db_tool import create_idea_sorted_table
 from src.a18_etl_toolbox.transformers import (
     etl_yell_raw_db_to_yell_agg_events_db,
     etl_yell_agg_events_db_to_yell_valid_events_db,
