+++ src/a18_etl_toolbox/test_tran/test_yell_agg_to_pidgin_raw.py
@@ -21,13 +21,15 @@ from src.a16_pidgin_logic._utils.str_a16 import (
     otx_label_str,
     unknown_word_str,
 )
-from src.a17_idea_logic._utils.str_a17 import idea_number_str
+from src.a17_idea_logic._utils.str_a17 import (
+    idea_number_str,
+    yell_agg_str,
+    yell_valid_str,
+)
 from src.a17_idea_logic.idea_db_tool import (
     upsert_sheet,
     sheet_exists,
     _get_pidgen_idea_format_filenames,
-    yell_agg_str,
-    yell_valid_str,
     get_default_sorted_list,
 )
 from src.a18_etl_toolbox.tran_path import create_yell_pidgin_path
