+++ src/a19_world_logic/test_00_sound_to_yell/test_world_etl01_yell_raw_yell_agg.py
@@ -2,13 +2,8 @@ from src.a00_data_toolbox.file_toolbox import create_path
 from src.a02_finance_logic._utils.strs_a02 import fisc_tag_str
 from src.a06_bud_logic._utils.str_a06 import face_name_str, event_int_str
 from src.a15_fisc_logic._utils.str_a15 import cumlative_minute_str, hour_tag_str
-from src.a17_idea_logic.idea_db_tool import (
-    get_sheet_names,
-    upsert_sheet,
-    yell_raw_str,
-    yell_agg_str,
-    sheet_exists,
-)
+from src.a17_idea_logic._utils.str_a17 import yell_agg_str, yell_raw_str
+from src.a17_idea_logic.idea_db_tool import get_sheet_names, upsert_sheet, sheet_exists
 from src.a19_world_logic.world import worldunit_shop
 from src.a19_world_logic._utils.env_a19 import (
     get_module_temp_dir as worlds_dir,
